Modified bootstrap-datepicker.js' formatDate method. This was extracting UTC values for an input date like getUTCDate(), getUTCDay(), getUTCMonth() etc. Changed them to their respective getDate(), getDay(), getMonth() etc.
Using UTC values is incorrect because input date is not bound to be UTC date. So if I pass a UTC +0530 date to this function, for the first 5:30 hours (0000 am to 0530am) it returns the previous day in the MM/dd/yyyy format. This scenario is more obvious when used with data-binding scenarios where data from an existing data source is bound to the input that's the target for the date-picker.